rename repository, fix todoRepository
